--- react-materials/scaffolds/ice-application-management-admin/src/components/BasicNotFound/index.js
@@ -1,3 +1,0 @@
-import BasicNotFound from './BasicNotFound';
-
-export default BasicNotFound;
