+++ react-materials/scaffolds/ice-application-management-admin/src/layouts/BasicLayout/components/Footer/index.js
