+++ react-materials/scaffolds/ice-application-management-admin/src/layouts/BasicLayout/components/Logo/Logo.jsx
